feat: file metrics collector rock
Katib ROCKs are part of main epic for building secure images using ROCKs.

ROCK for Katib file-metrics-collector is part of katib-config.
Based on https://github.com/canonical/katib-rocks/pull/5

Since configuration is provided when this container is ran in the pod it is expected to fail when running manually. Howvever, it is enough to ensure that service is being started.

Summary of changes:
- Initial version of TFEvent metrics collector ROCK.