feat: suggestion-hyperopt rock
Katib ROCKs are part of main epic for building secure images using ROCKs.

ROCK for Katib suggestion-hyperopt is part of katib-config.

Since configuration is provided when this container is ran in the pod it is expected to fail when running manually. Howvever, it is enough to ensure that service is being started.

Summary of changes:
- Initial version of suggestion-hyperopt ROCK.