Feat earlystopping medianstop rock (#12)
* feat: earlystopping-medianstop rock
Katib ROCKs are part of main epic for building secure images using ROCKs.

ROCK for Katib earlystopping-medianstop is part of katib-config.

Based on #1

Summary of changes:
- Created Katib earlystopping-medianstop ROCK using best practices.
- Added workdir to service and modified command.
- Removed not needed override-build.
Since configuration is provided when this container is ran in the pod it is expected to fail when running manually. However, it is enough to ensure that service is being started.